backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.37
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
better-exceptions==0.3.3
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.12.5